requirements.txt
msal
requests
beautifulsoup4
python-dotenv
fuzzywuzzy
